backend/requirements.txt
# FastAPI Backend Dependencies
fastapi
uvicorn[standard]
python-multipart
pydantic[email]
pillow
numpy
opencv-python-headless
requests
tensorflow
matplotlib
scipy
reportlab

# Database
sqlalchemy
psycopg2-binary

# Authentication
python-jose[cryptography]

# Environment
python-dotenv
